Add backstop for 2025 template failures
